excel-pdf-values-validator/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
asyncpg==0.29.0
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
PyMuPDF==1.23.8
openpyxl==3.1.2
pandas==2.1.3
numpy>=1.24.0,<2.0
transformers>=4.30.0
huggingface-hub>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
pgvector==0.2.4
python-dotenv==1.0.0
# Multimodal model dependencies
Pillow==10.1.0
pytesseract==0.3.10

# Celery and distributed task processing
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0
kombu==5.3.4

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
requests==2.31.0
psutil==5.9.6

# Monitoring and metrics
prometheus-client==0.19.0

# Quality gates
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
types-redis==4.6.0.8
types-requests==2.31.0.10
